left join 和權限問題

+++ controllers/post.js
@@ -35,8 +35,7 @@ module.exports = {
     queryData.offset = parseInt(req.query.offset, 10)
     queryData.limit = parseInt(req.query.limit, 10)
     queryData.order = 'createdAt'
-    let result = await getUnpublishedPosts(false, queryData)
-    if (!result) return res.json(errorMessage.fetchFail)
+    let result = await getPublishedPosts(null, queryData)
     return res.json(result)
   },
   getPostsByPlaceId: async (req, res) => {
@@ -58,64 +57,35 @@ module.exports = {
   getPosts: async (req, res) => {
     // 取得單一使用者的複數食記
     const checkedList = ['offset', 'limit', 'order']
-    const sessionId = parseInt(req.session.userId, 10)
-    const userId = parseInt(req.params.user_id, 10)
-    console.log(!req.session.user || req.session.userId != req.params.user_id)
+    const sessionId = req.session.userId || null
     if (!checkedList.every(key => Object.keys(req.query).includes(key))) {
       return res.json(errorMessage.missingParameter)
     }
+    if (!req.params.user_id) return res.json(errorMessage.missingParameter)
     const queryData = {}
     queryData.userId = parseInt(req.params.user_id, 10)
     queryData.offset = parseInt(req.query.offset, 10)
     queryData.limit = parseInt(req.query.limit, 10)
+    queryData.published = !JSON.parse(req.query.unpublished)
     queryData.order = req.query.order
-    let result = ""
-    if (!req.session.user || sessionId !== userId) {
-      result = await getUnpublishedPosts(false, queryData)
-      if (!result) return res.json(errorMessage.fetchFail)
-      return res.json(result)
-    }
-    if (!req.query.unpublished) {
-      result = await getUnpublishedPosts(false, queryData)
-    }
-    if (req.query.unpublished === 'true') {
-      result = await getUnpublishedPosts(true, queryData)
-    }
-    if (req.query.unpublished === 'false') {
-      result = await getUnpublishedPosts(false, queryData)
-    }
-    if (!result) return res.json(errorMessage.fetchFail)
+    let result
+    result = await getPublishedPosts(sessionId, queryData)
     return res.json(result)
   },
   getPost: async (req, res) => {
     // 取得單一使用者的單一食記
-    const sessionId = parseInt(req.session.userId, 10)
+    const sessionId = req.session.userId || null
     const userId = parseInt(req.query.user_id, 10)
     const postId = parseInt(req.params.post_id, 10)
-    if (!req.session.user || sessionId !== userId) {
-      let result = await getUnpublishedPost(false, postId)
-      if (!result) return res.json(errorMessage.fetchFail)
-      return res.json(result)
-    }
-    let result = await getUnpublishedPost(true, postId)
+    let result = await getPublishedPost(postId, sessionId)
     if (!result) return res.json(errorMessage.fetchFail)
     return res.json(result)
   },
   addPost: async (req, res) => {
-    const sessionId = parseInt(req.session.userId, 10)
-    const userId = parseInt(req.body.user_id, 10)
-    console.log(!req.session.user)
-    console.log(sessionId !== userId)
-    console.log(sessionId)
-    console.log(userId)
-
-    if (!req.session.user || sessionId !== userId) {
-      return res.json(errorMessage.unauthorized)
-    }
+    const userId = req.session.userId
     if (!req.files) return res.json(errorMessage.noPhotos)
     if (req.files.length < 1) return res.json(errorMessage.noPhotos)
     const checkedList = [
-      'user_id',
       'restaurant_id',
       'title',
       'content',
@@ -130,7 +100,6 @@ module.exports = {
     const imageResult = await uploadImage(req)
     if (!imageResult) return res.json(errorMessage.fetchFail)
     const {
-      user_id,
       restaurant_id,
       title,
       content,
@@ -140,7 +109,7 @@ module.exports = {
     let postResult = null
     try {
       postResult = await PostDb.create({
-        user_id,
+        user_id: userId,
         restaurant_id,
         title,
         content,
@@ -183,15 +152,10 @@ module.exports = {
     return res.json(result)
   },
   editPost: async (req, res) => {
-    const sessionId = parseInt(req.session.userId, 10)
-    const userId = parseInt(req.body.user_id, 10)
-    if (!req.session.userId || sessionId !== userId) {
-      return res.json(errorMessage.unauthorized)
-    }
+    const sessionId = req.session.userId
     const imageCount = req.files.length
     const postId = parseInt(req.params.post_id, 10)
     const imageResult = await uploadImage(req)
-    console.log(imageResult)
     if (!imageResult) return res.json(errorMessage.fetchFail)
     const {
       restaurant_id,
@@ -211,14 +175,14 @@ module.exports = {
           is_published,
         },
         {
-          where: { id: postId },
+          where: { id: postId, user_id: sessionId },
         }
       )
     } catch (err) {
       console.log(err)
       return res.json(errorMessage.fetchFail)
     }
-    console.log(editResult)
+    if (editResult[0] === 0) return res.json(errorMessage.postNotFound)
     try {
       await PictureDb.destroy({
         where: {
@@ -244,7 +208,7 @@ module.exports = {
     return res.json(okMessage)
   },
   deletePost: async (req, res) => {
-    const sessionId = parseInt(req.session.userId, 10)
+    const sessionId = req.session.userId
     const postId = parseInt(req.params.post_id, 10)
     let result = null
     try {
@@ -306,18 +270,20 @@ async function uploadImage(req) {
   return imgurResultArr
 }
 
-async function getUnpublishedPosts(unpublished = false, queryData) {
-  const { userId, offset, limit, order } = queryData
-  let where = { user_id: userId, is_deleted: false, is_published: true }
-  if (!userId) {
-    where = { is_deleted: false, is_published: true }
-  }
-  if (unpublished && userId) {
-    where = { user_id: userId, is_deleted: false, is_published: false }
-  }
+async function getPublishedPosts(sessionId, queryData) {
+  const {
+    userId = null,
+    offset = 0,
+    limit = 5,
+    order = 'createdAt',
+    published = true,
+  } = queryData
+  if (sessionId !== userId && published === false)
+    return errorMessage.unauthorized
+  let where = null
+  if (!userId) where = { is_deleted: false, is_published: published }
+  else where = { user_id: userId, is_deleted: false, is_published: published }
   let result = null
-  let imageResult = null
-  let imageArr = []
   try {
     result = await PostDb.findAndCountAll({
       where,
@@ -325,45 +291,22 @@ async function getUnpublishedPosts(unpublished = false, queryData) {
       order: [[order, 'DESC']],
       offset,
       limit,
+      include: {
+        model: PictureDb,
+        attributes: ['food_picture_url'],
+      },
     })
-    console.log(result)
   } catch (err) {
-    return false
+    console.log(err)
+    return errorMessage.postNotFound
   }
-  if (!result) return {}
-  for (let i = 0; i < result.rows.length; i++) {
-    try {
-      imageResult = await PictureDb.findOne({
-        where: { post_id: result.rows[i].id },
-      })
-      console.log(imageResult)
-    } catch (err) {
-      console.log(err)
-      return false
-    }
-    if (!imageResult) continue
-    imageArr.push({
-      postId: imageResult.post_id,
-      link: imageResult.food_picture_url,
-    })
-  }
-  const data = {
-    postCounts: result.count,
-    posts: result.rows.map(each => each.dataValues),
-    images: imageArr,
-  }
-  console.log(data)
-  return data
+  if (!result) return errorMessage.postNotFound
+  return result
 }
 
-async function getUnpublishedPost(unpublished = false, postId) {
-  let where = { id: postId, is_deleted: false, is_published: true }
-  if (unpublished) {
-    where = { id: postId, is_deleted: false }
-  }
+async function getPublishedPost(postId, sessionId) {
+  let where = { id: postId, is_deleted: false }
   let result = null
-  let imageResult = null
-  let imageArr = []
   try {
     result = await PostDb.findOne({
       where,
@@ -376,26 +319,21 @@ async function getUnpublishedPost(unpublished = false, postId) {
         'visited_time',
         'is_published',
       ],
+      include: {
+        model: PictureDb,
+        attributes: ['food_picture_url'],
+      },
     })
   } catch (err) {
     console.log(err)
     return false
   }
-  if (!result) return {}
-  try {
-    imageResult = await PictureDb.findAndCountAll({
-      where: { post_id: result.id },
-    })
-  } catch (err) {
-    console.log(err)
-    return false
-  }
-  imageResult.rows.forEach(each => imageArr.push(each.food_picture_url))
-  const data = {
-    post: result,
-    images: imageArr,
-  }
-  return data
+  if (!result) return errorMessage.postNotFound
+  console.log(result)
+  if (result.is_published === false && result.user_id !== sessionId)
+    return errorMessage.unauthorized
+
+  return result
 }
 
 async function getPostsByPlaceId(unpublished = false, queryData) {
